requirements.txt
asgiref==3.2.3
Babel==2.8.0
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-heroku==0.3.1
django-phonenumber-field==4.0.0
django-summernote==0.8.11.6
djangorestframework==3.11.0
gunicorn==20.0.4
phonenumbers==8.11.4
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
